+++ Acha/Acha/Assets.xcassets/crazy.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "이미지 추가.png",
+      "filename" : "crazy.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
