+++ Acha/Acha/Assets.xcassets/defaultSelectImage.imageset/Contents.json
@@ -1,7 +1,7 @@
 {
   "images" : [
     {
-      "filename" : "crazy.png",
+      "filename" : "defaultSelectImage.png",
       "idiom" : "universal",
       "scale" : "1x"
     },
