--- Acha/Acha/Assets.xcassets/defaultSelectImage.imageset/Contents.json~HEAD
@@ -1,21 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "이미지 추가 1.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
