--- Acha/Acha/Data/CommunityDTO.swift
@@ -1,12 +1,0 @@
-//
-//  CommunityDTO.swift
-//  Acha
-//
-//  Created by hong on 2022/11/28.
-//
-
-import Foundation
-
-struct CommunityDTO: Codable {
-    let posts: [PostDTO]
-}
