+++ Acha/Acha/Util/Extension/String+.swift
@@ -5,7 +5,7 @@
 //  Created by 조승기 on 2022/11/21.
 //
 
-import Foundation
+import UIKit
 
 extension String {
     func convertToDateFormat(format: String) -> Date {
@@ -16,5 +16,34 @@ extension String {
             return date
         }
         return Date()
+    }
+    
+    func stringCheck(pattern: String) -> Bool {
+        do {
+            let regex = try NSRegularExpression(pattern: pattern, options: .caseInsensitive)
+            let range = NSRange(location: 0, length: self.count)
+            if regex.firstMatch(in: self, range: range) != nil {
+                return true
+            } else {
+                return false
+            }
+        } catch {
+            print(error.localizedDescription)
+            return false
+        }
+    }
+    
+    func generateQRCode() -> UIImage? {
+        let data = self.data(using: String.Encoding.ascii)
+
+        if let filter = CIFilter(name: "CIQRCodeGenerator") {
+            filter.setValue(data, forKey: "inputMessage")
+            let transform = CGAffineTransform(scaleX: 3, y: 3)
+
+            if let output = filter.outputImage?.transformed(by: transform) {
+                return UIImage(ciImage: output)
+            }
+        }
+        return nil
     }
 }
