--- Acha/Acha/Util/Extension/String+Extension.swift
@@ -1,40 +1,0 @@
-//
-//  String+Extension.swift
-//  Acha
-//
-//  Created by hong on 2022/11/22.
-//
-
-import Foundation
-import UIKit
-
-extension String {
-    func stringCheck(pattern: String) -> Bool {
-        do {
-            let regex = try NSRegularExpression(pattern: pattern, options: .caseInsensitive)
-            let range = NSRange(location: 0, length: self.count)
-            if regex.firstMatch(in: self, range: range) != nil {
-                return true
-            } else {
-                return false
-            }
-        } catch {
-            print(error.localizedDescription)
-            return false
-        }
-    }
-    
-    func generateQRCode() -> UIImage? {
-        let data = self.data(using: String.Encoding.ascii)
-
-        if let filter = CIFilter(name: "CIQRCodeGenerator") {
-            filter.setValue(data, forKey: "inputMessage")
-            let transform = CGAffineTransform(scaleX: 3, y: 3)
-
-            if let output = filter.outputImage?.transformed(by: transform) {
-                return UIImage(ciImage: output)
-            }
-        }
-        return nil
-    }
-}
