--- Acha/FBProviderType.swift
@@ -1,8 +1,0 @@
-//
-//  FBProviderType.swift
-//  Acha
-//
-//  Created by 조승기 on 2022/11/30.
-//
-
-import Foundation
